+++ test/integration/analyzers/nodejs_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/fossas/fossa-cli/files"
 	"github.com/fossas/fossa-cli/module"
 	"github.com/fossas/fossa-cli/pkg"
-	"github.com/fossas/fossa-cli/tests/project_initializer/node"
-	"github.com/fossas/fossa-cli/tests/testtools"
+	"github.com/fossas/fossa-cli/test/project_initializer/node"
+	"github.com/fossas/fossa-cli/test/testtools"
 	"github.com/stretchr/testify/assert"
 )
